Add Qt 6 build support, panelizer v2, and Gerber preview
Qt 6 port: the tree now builds and runs against Qt 6 (6.5+) as well as
Qt 5.15 from a single source, selected by whichever qmake is used. All
removed/changed APIs ported to version-agnostic equivalents where one
exists, with QT_VERSION guards only where unavoidable (QMatrix->QTransform,
qrand->QRandomGenerator, setPaperSize/paperRect->QPageSize, QFlags arg
casts, QStandardPaths::AppLocalDataLocation, setContent ParseResult,
QTextCodec/setCodec guards via core5compat). phoenix.pro pulls in
core5compat and svgwidgets under Qt 6.

Panelizer v2: full File->Export->for Production->Panelize wizard that
produces real RS-274X Gerbers + Excellon drill, composite Edge_Cuts
outline, frame, and V-cut/mouse-bite separations. Adds vendor panel
presets, auto-fit panel mode, heavy-load confirmation, progress dialog,
and a dedicated panel board item.

Gerber preview: standalone GerberPreviewDialog opens after any 274X
export (single-board or panelized) so artifacts can be verified in-app
before shipping to a fab house.

Adds a reference Qt 6 container build (docker/Dockerfile.qt6,
docker/build-linux-qt6.sh) plus QtTest coverage (test_gerber,
test_gerberpreview, test_panelizer). Some polish still pending.

Next: an interactive part design editor to fill a long-standing gap in
the program.

+++ src/items/itembase.cpp
@@ -1873,7 +1873,8 @@ void ItemBase::debugInfo2(const QString & msg) const
 	                   .arg(this->instanceTitle())
 	                   .arg(this->viewLayerID())
 	                   .arg(this->viewLayerPlacement())
-	                   .arg(this->wireFlags())
+	                   // NOTE: Qt 6 removed the implicit QFlags->int conversion in QString::arg().
+	                   .arg(static_cast<int>(this->wireFlags()))
 			   .arg((long) dynamic_cast<const QGraphicsItem *>(this), 0, 16)
 	                   .arg(m_viewID)
 	                   .arg(this->zValue())
